requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
websockets==12.0
python-multipart==0.0.6
python-dotenv==1.0.0
deepgram-sdk==2.12.0
google-generativeai==0.3.1
elevenlabs==0.2.26
pdfplumber==0.10.3
webrtcvad==2.0.10
aiofiles==23.2.1
groq
numpy